Initial commit of rescuetime plugin for mothermayi
I haven't tested this yet because I'm on a plane, but I'm pretty sure it
will work based on reading the official recuetime plugin I downloaded a
while ago. We just make a request using the API key for recuetime we
pull from the mothermayi config and that should do it.