Handle exceptions due to disconnect
This makes it so that the hook will do the correct thing on...say...a
plane. It *is* a bit sad that the disconnection means that we throw away
the commit data. But what are you going to do? Store stuff in some kind
of tiny database and push all of it up next time a commit happens...?